requirements.txt
discord.py==2.2.0
dnspython==2.2.1
PyNaCl==1.5.0
async-timeout==3.0.0
youtube-search-python==1.6.6
asyncio==3.4.3
mutagen==1.45.1
requests==2.28.1
beautifulsoup4==4.11.1
urllib3==1.26.12
pygame==2.1.2
lxml==4.9.1
yt_dlp==2023.2.17